@@ -1,45 +1,61 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 36
     defaultConfig {
         applicationId "com.samuelberrien.phyvr"
-        minSdkVersion 19
-        targetSdkVersion 28
+        minSdkVersion 30
+        targetSdkVersion 36
         versionCode 17
-        versionName "1.5g"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        versionName "2.0"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
         externalNativeBuild {
             cmake {
-                arguments "-DCMAKE_BUILD_TYPE=release"
+                arguments "-DCMAKE_VERBOSE_MAKEFILE=ON"
             }
+        }
+        ndk {
+            //noinspection ChromeOsAbiSupport
+            abiFilters "arm64-v8a"
         }
     }
     buildTypes {
+        debug {
+            externalNativeBuild {
+                cmake {
+                    cppFlags "-O1"
+                }
+            }
+        }
         release {
-            minifyEnabled true
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-            proguardFiles.add(file('../../proguard-gvr.txt'))
-
+            externalNativeBuild {
+                cmake {
+                    cppFlags "-O3"
+                }
+            }
+            minifyEnabled false
+            signingConfig signingConfigs.debug
         }
     }
     externalNativeBuild {
         cmake {
             path "CMakeLists.txt"
+            version "3.30.3"
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_1_10
+        targetCompatibility JavaVersion.VERSION_1_10
     }
-    ndkVersion '21.0.6113669'
+    ndkVersion "27.3.13750724"
+    namespace 'com.samuelberrien.phyvr'
 }
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-    implementation 'com.google.vr:sdk-base:1.170.0'
+    implementation 'androidx.appcompat:appcompat:1.7.1'
+    testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.3.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.7.0'
+    implementation platform("org.jetbrains.kotlin:kotlin-bom:1.8.0")
 }
